rename: 모달 디렉토리 변경
- NotificationModal 모달 src/modal/ 로 이동

#64
